Fixes and refactoring regarding champion.json
As the champion.json file was deleted, errors would be raised, if the champion.json file was not there initially.
With this fix, we moved some functions around to not have circular imports.
Furthermore, champion.json is now updated right before it is used, instead of being a global dict in the modules that use it.