Removed old direcotories and files and fixed critical bug
Removed the ocr pictures, which are really not supposed to be in the repo.
Also removed the "./data/strings" directory, as this is a depricated directory as far as im aware.
Fixed a bug in the riot commands module created by fbe355155217aac50a2a677287d946e8dee68721.
Also updated .gitignore to be less messy and reflect the current situation better.