Proposal to fix issue #143
To fix the problem of writing futures to a file, we will set the values of all
futures to None before writing them to a file.
This will be done through a list of attributes, which has to include all attributes of Guild State that are not pickleable.
Afterwards, the values get restored to their original again.