Proposal of channel command rework
With this rework, we can get rid of the old loop in cog tasks, that checks if the temporary channels can be deleted and we also get rid of the `guild_state.tmp_channel_ids` attribute.
We trigger a coroutine on the created temporary channel, that deletes the channel after a given time.
The implementation for this is conceptually the same as the discord.py solution for deleting messages after a delay. Additonally, we utilize the cooldown decorator from discord.py to set a cooldown for each user and each guild to avoid spam.
Deleting temporary channel handlers in events.py and now unused function in timers.py.