Closes #149
If a summoner has not played any ranked games, their rank_values dict is set to `{}`.
For some reason, if this is the case, the summoner is not removed in the for loop and if iterating over the for loop and printing `summoner.name` without any condition, the summoner is also excluded.
Yet, when using a list comprehension the summoner can be detected and discarded.